AI & Result/requirements.txt
Flask
google-cloud-language
reportlab